update eBQP to 0.0.1

+++ examples/2D_robot/models.py
@@ -7,7 +7,7 @@ from qpth.qp import QPFunction
 import numpy as np
 from my_classes import test_solver as solver
 import time
-from eBQP import eBQP
+from eBQP import eBQP, eBQP_I, eBQP_g
 
 
 class BarrierNet(nn.Module):
@@ -274,7 +274,7 @@ class ABNet(nn.Module):
         G0 = torch.cat([G10, G20],dim=1)
         h0 = torch.cat([h1, h2],dim=1)
 
-        H = Variable(torch.eye(self.nCls))
+        H = Variable(torch.eye(self.nCls,dtype=torch.float64))
         H = H.unsqueeze(0).expand(nBatch, self.nCls, self.nCls).to(self.device)
         
         x = eBQP(H, -x31, G0, h0)
